# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eleicoes2026 API",
    "description": "Dados e indicadores eleitorais (somente leitura).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/candidatos": {
      "get": {
        "tags": [
          "candidatos"
        ],
        "summary": "Candidatos de um grupo",
        "description": "Votos, % dos válidos, penetração, resultado e partido, ordenados por votos.",
        "operationId": "candidatos_api_candidatos_get",
        "parameters": [
          {
            "name": "grupo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id do grupo em config/grupos.yaml (ex.: missao_2026).",
              "title": "Grupo"
            },
            "description": "Id do grupo em config/grupos.yaml (ex.: missao_2026)."
          },
          {
            "name": "uf",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UF"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uf"
            }
          },
          {
            "name": "cargo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Cargo"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cargo"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 200,
              "title": "Limite"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListaCandidatos"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidatos/{ano}/{sq_candidato}": {
      "get": {
        "tags": [
          "candidatos"
        ],
        "summary": "Ficha do candidato",
        "description": "Votos por UF e município (top N), gastos, receitas por fonte e custo por voto.",
        "operationId": "ficha_api_candidatos__ano___sq_candidato__get",
        "parameters": [
          {
            "name": "ano",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Ano"
            }
          },
          {
            "name": "sq_candidato",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sq Candidato"
            }
          },
          {
            "name": "top",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Municípios no ranking.",
              "default": 10,
              "title": "Top"
            },
            "description": "Municípios no ranking."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FichaCandidato"
                }
              }
            }
          },
          "404": {
            "description": "Candidato inexistente."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mapa": {
      "get": {
        "tags": [
          "mapa"
        ],
        "summary": "Indicador por território",
        "description": "`valores` por território (município, `município-zona` ou célula H3) + escala sugerida.",
        "operationId": "mapa_api_mapa_get",
        "parameters": [
          {
            "name": "ano",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Ano"
            }
          },
          {
            "name": "cargo",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Cargo"
            }
          },
          {
            "name": "uf",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UF"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uf"
            }
          },
          {
            "name": "nivel",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Nivel",
              "default": "municipio"
            }
          },
          {
            "name": "grupo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Id do grupo; exclusivo com sq_candidato.",
              "title": "Grupo"
            },
            "description": "Id do grupo; exclusivo com sq_candidato."
          },
          {
            "name": "sq_candidato",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "SQ_CANDIDATO; exclusivo com grupo.",
              "title": "Sq Candidato"
            },
            "description": "SQ_CANDIDATO; exclusivo com grupo."
          },
          {
            "name": "indicador",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Indicador",
              "default": "penetracao"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mapa"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mapa/pontos": {
      "get": {
        "tags": [
          "mapa"
        ],
        "summary": "Locais de votação (densidade)",
        "description": "`{lat, lon, votos}` por local, mais votados primeiro; paginado por UF.",
        "operationId": "pontos_api_mapa_pontos_get",
        "parameters": [
          {
            "name": "ano",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Ano"
            }
          },
          {
            "name": "cargo",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Cargo"
            }
          },
          {
            "name": "uf",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UF"
            }
          },
          {
            "name": "grupo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Id do grupo; exclusivo com sq_candidato.",
              "title": "Grupo"
            },
            "description": "Id do grupo; exclusivo com sq_candidato."
          },
          {
            "name": "sq_candidato",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "SQ_CANDIDATO; exclusivo com grupo.",
              "title": "Sq Candidato"
            },
            "description": "SQ_CANDIDATO; exclusivo com grupo."
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50000,
              "minimum": 1,
              "default": 5000,
              "title": "Limite"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pontos"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gastos": {
      "get": {
        "tags": [
          "gastos"
        ],
        "summary": "Gastos e receitas do grupo",
        "description": "Custo por voto (contratado/pago), receita por fonte, % público e % autofinanciamento.",
        "operationId": "gastos_api_gastos_get",
        "parameters": [
          {
            "name": "grupo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id do grupo em config/grupos.yaml (ex.: missao_2026).",
              "title": "Grupo"
            },
            "description": "Id do grupo em config/grupos.yaml (ex.: missao_2026)."
          },
          {
            "name": "uf",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UF"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uf"
            }
          },
          {
            "name": "cargo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Cargo"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cargo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gastos"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/comparativo": {
      "get": {
        "tags": [
          "comparativo"
        ],
        "summary": "Evolução 2022→2026",
        "description": "Δ penetração (‰), swing (p.p.), retenção e ganho por AMC, mais KPIs do recorte.",
        "operationId": "comparativo_api_comparativo_get",
        "parameters": [
          {
            "name": "comparacao",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id em `comparacoes` (ex.: evolucao_mbl).",
              "title": "Comparacao"
            },
            "description": "Id em `comparacoes` (ex.: evolucao_mbl)."
          },
          {
            "name": "cargo",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Cargo"
            }
          },
          {
            "name": "uf",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UF"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uf"
            }
          },
          {
            "name": "mesmos_candidatos",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Só pessoas que concorreram nos dois anos (pessoa_id).",
              "default": false,
              "title": "Mesmos Candidatos"
            },
            "description": "Só pessoas que concorreram nos dois anos (pessoa_id)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comparativo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/municipios/{cd_mun_ibge}": {
      "get": {
        "tags": [
          "municipios"
        ],
        "summary": "Resumo do município",
        "description": "Desempenho de cada grupo no município, por cargo.",
        "operationId": "municipio_api_municipios__cd_mun_ibge__get",
        "parameters": [
          {
            "name": "cd_mun_ibge",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{7}$",
              "description": "Código IBGE de 7 dígitos.",
              "title": "Cd Mun Ibge"
            },
            "description": "Código IBGE de 7 dígitos."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoMunicipio"
                }
              }
            }
          },
          "404": {
            "description": "Município inexistente."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Ano": {
        "type": "integer",
        "enum": [
          2022,
          2026
        ],
        "title": "Ano",
        "description": "Anos de eleição geral suportados."
      },
...
      "Cargo": {
        "type": "string",
        "enum": [
          "PRESIDENTE",
          "GOVERNADOR",
          "SENADOR",
          "DEPUTADO FEDERAL",
          "DEPUTADO ESTADUAL",
          "DEPUTADO DISTRITAL"
        ],
        "title": "Cargo",
        "description": "`ds_cargo` como gravado nos Parquet (maiúsculas; o ETL normaliza)."
      },
...
      "FichaCandidato": {
        "properties": {
          "candidato": {
            "$ref": "#/components/schemas/CandidatoResumo"
          },
          "votos_total": {
            "type": "integer",
            "title": "Votos Total"
          },
          "votos_por_uf": {
            "items": {
              "$ref": "#/components/schemas/VotosUF"
            },
            "type": "array",
            "title": "Votos Por Uf"
          },
          "votos_por_municipio": {
            "items": {
              "$ref": "#/components/schemas/VotosMunicipio"
            },
            "type": "array",
            "title": "Votos Por Municipio",
            "description": "Top N por votos."
          },
          "gastos": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResumoCustoCandidato"
              },
              {
                "type": "null"
              }
            ],
            "description": "null se não há prestação de contas."
          },
          "receitas": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResumoReceitasOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "contas_parciais": {
            "type": "boolean",
            "title": "Contas Parciais",
            "description": "Prestação de contas ainda parcial (2026)."
          },
          "base_ipca": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Ipca",
            "description": "Mês-base da correção do IPCA (valores de 2022)."
          },
          "dt_geracao": {
            "type": "string",
            "title": "Dt Geracao"
          }
        },
        "type": "object",
        "required": [
          "candidato",
          "votos_total",
          "votos_por_uf",
          "votos_por_municipio",
          "gastos",
          "receitas",
          "contas_parciais",
          "base_ipca",
          "dt_geracao"
        ],
        "title": "FichaCandidato",
        "description": "Ficha completa: votação, geografia e finanças."
      },
...
      "Gastos": {
        "properties": {
          "grupo": {
            "type": "string",
            "title": "Grupo"
          },
          "ano": {
            "type": "integer",
            "title": "Ano"
          },
          "cargo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cargo"
          },
          "uf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uf"
          },
          "agregado": {
            "$ref": "#/components/schemas/ResumoCustoGrupo"
          },
          "receitas": {
            "$ref": "#/components/schemas/ResumoReceitasOut"
          },
          "por_candidato": {
            "items": {
              "$ref": "#/components/schemas/GastoCandidato"
            },
            "type": "array",
            "title": "Por Candidato"
          },
          "contas_parciais": {
            "type": "boolean",
            "title": "Contas Parciais"
          },
          "base_ipca": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Ipca"
          },
          "dt_geracao": {
            "type": "string",
            "title": "Dt Geracao"
          }
        },
        "type": "object",
        "required": [
          "grupo",
          "ano",
          "cargo",
          "uf",
          "agregado",
          "receitas",
          "por_candidato",
          "contas_parciais",
          "base_ipca",
          "dt_geracao"
        ],
        "title": "Gastos",
        "description": "Corpo de GET /gastos.",
        "examples": [
          {
            "agregado": {
              "candidatos_sem_voto_excluidos": 0,
              "custo_voto_contratado": 88.98,
              "custo_voto_pago": 72.03,
              "despesa_contratada": 105000.0,
              "despesa_paga": 85000.0,
              "divida": 20000.0,
              "mediana_custo_voto_contratado": 63.89,
              "votos": 1180
            },
            "ano": 2026,
            "cargo": "DEPUTADO FEDERAL",
            "contas_parciais": true,
            "dt_geracao": "2026-10-06",
            "grupo": "missao_2026",
            "por_candidato": [],
            "receitas": {
              "pct_autofinanciamento": 8.0,
              "pct_publico": 64.0,
              "por_categoria": {
                "fefc": 70000.0,
                "fundo_partidario": 10000.0
              },
              "receita_financeira": 123000.0,
              "receita_total": 125000.0
            },
            "uf": "SP"
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Indicador": {
        "type": "string",
        "enum": [
          "penetracao",
          "pct_validos",
          "votos"
        ],
        "title": "Indicador",
        "description": "Indicadores mapeáveis (spec §2): taxas em coroplético, absoluto em símbolo."
      },
...
      "ListaCandidatos": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limite": {
            "type": "integer",
            "title": "Limite"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "itens": {
            "items": {
              "$ref": "#/components/schemas/CandidatoResumo"
            },
            "type": "array",
            "title": "Itens"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limite",
          "offset",
          "itens"
        ],
        "title": "ListaCandidatos",
        "description": "Página de candidatos.",
        "examples": [
          {
            "itens": [
              {
                "ano": 2026,
                "cargo": "DEPUTADO FEDERAL",
                "nm_urna": "A",
                "partido": {
                  "numero": 14,
                  "sigla": "MISSÃO"
                },
                "pct_validos": 8.06,
                "penetracao": 57.14,
                "resultado": "SUPLENTE",
                "sg_uf": "SP",
                "situacao": "APTO",
                "sq_candidato": 3,
                "votos": 1000
              }
            ],
            "limite": 200,
            "offset": 0,
            "total": 1
          }
        ]
      },
      "Mapa": {
        "properties": {
          "ano": {
            "type": "integer",
            "title": "Ano"
          },
          "nivel": {
            "type": "string",
            "title": "Nivel"
          },
          "indicador": {
            "type": "string",
            "title": "Indicador"
          },
          "valores": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Valores"
          },
          "detalhes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Detalhe"
            },
            "type": "object",
            "title": "Detalhes"
          },
          "escala_sugerida": {
            "$ref": "#/components/schemas/EscalaSugerida"
          },
          "unidade": {
            "type": "string",
            "title": "Unidade"
          },
          "denominador": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Denominador"
          },
          "n_candidaturas": {
            "type": "integer",
            "title": "N Candidaturas"
          },
          "dt_geracao": {
            "type": "string",
            "title": "Dt Geracao"
          }
        },
        "type": "object",
        "required": [
          "ano",
          "nivel",
          "indicador",
          "valores",
          "detalhes",
          "escala_sugerida",
          "unidade",
          "denominador",
          "n_candidaturas",
          "dt_geracao"
        ],
        "title": "Mapa",
        "description": "Corpo de GET /mapa. Chaves: IBGE (município), `IBGE-zona`, ou índice H3.",
        "examples": [
          {
            "ano": 2026,
            "denominador": "aptos",
            "detalhes": {
              "3550308": {
                "aptos": 15000,
                "taxa": 73.33,
                "validos": 10500,
                "votos": 1100
              }
            },
            "dt_geracao": "2026-10-06",
            "escala_sugerida": {
              "paleta": "viridis",
              "quebras": [
                20.0,
                40.0,
                60.0,
                70.0
              ],
              "tipo": "sequencial"
            },
            "indicador": "penetracao",
            "n_candidaturas": 2,
            "nivel": "municipio",
            "unidade": "‰",
            "valores": {
              "3509502": 20.0,
              "3550308": 73.33
            }
          }
        ]
      },
...
      "Nivel": {
        "type": "string",
        "enum": [
          "municipio",
          "zona",
          "h3"
        ],
        "title": "Nivel",
        "description": "Recorte espacial do mapa."
      },
...
      "Pontos": {
        "properties": {
          "ano": {
            "type": "integer",
            "title": "Ano"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Locais com voto no recorte."
          },
          "limite": {
            "type": "integer",
            "title": "Limite"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "truncado": {
            "type": "boolean",
            "title": "Truncado",
            "description": "Há mais páginas além desta."
          },
          "pontos": {
            "items": {
              "$ref": "#/components/schemas/Ponto"
            },
            "type": "array",
            "title": "Pontos"
          },
          "dt_geracao": {
            "type": "string",
            "title": "Dt Geracao"
          }
        },
        "type": "object",
        "required": [
          "ano",
          "total",
          "limite",
          "offset",
          "truncado",
          "pontos",
          "dt_geracao"
        ],
        "title": "Pontos",
        "description": "Corpo de GET /mapa/pontos (densidade de votos por local)."
      },
...
      "UF": {
        "type": "string",
        "enum": [
          "AC",
          "AL",
          "AM",
          "AP",
          "BA",
          "CE",
          "DF",
          "ES",
          "GO",
          "MA",
          "MG",
          "MS",
          "MT",
          "PA",
          "PB",
          "PE",
          "PI",
          "PR",
          "RJ",
          "RN",
          "RO",
          "RR",
          "RS",
          "SC",
          "SE",
          "SP",
          "TO"
        ],
        "title": "UF",
        "description": "Unidades da federação (sem exterior: `ZZ` não tem mapa)."
      },
...
      "VotosUF": {
        "properties": {
          "uf": {
            "type": "string",
            "title": "Uf"
          },
          "votos": {
            "type": "integer",
            "title": "Votos"
          }
        },
        "type": "object",
        "required": [
          "uf",
          "votos"
        ],
        "title": "VotosUF",
        "description": "Votos do candidato numa UF."
      }
    }
  }
}